+++ last-order-api/src/main/java/com/flab/store/controller/StoreController.java
@@ -1,14 +1,10 @@
 package com.flab.store.controller;
 
-import com.flab.store.domain.Store;
 import com.flab.store.dto.request.AddStoreRequest;
 import com.flab.store.dto.request.UpdateStoreRequest;
 import com.flab.store.dto.response.StoreResponse;
 import com.flab.store.service.StoreService;
 import lombok.RequiredArgsConstructor;
-import org.hibernate.sql.Update;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 import javax.validation.Valid;
@@ -21,38 +17,38 @@ public class StoreController {
 
     private final StoreService storeService;
 
-    @GetMapping("/searchStoreInfo")
-    public Store searchStoreInfo() {
-        return storeService.searchStoreInfo();
+    @GetMapping("/searchStoreInfo/{storeId}")
+    public StoreResponse searchStoreInfo(@PathVariable Long storeId) {
+        return storeService.searchStoreInfo(storeId);
     }
 
-    @GetMapping("/searchAllMyStore")
-    public List<Store> searchAllMyStore() {
-        return storeService.searchAllMyStore();
+    @GetMapping("/searchAllMyStore/{ownerId}")
+    public List<StoreResponse> searchAllMyStore(@PathVariable Long ownerId) {
+        return storeService.searchAllMyStore(ownerId);
     }
 
     @PostMapping("/add")
-    public ResponseEntity<StoreResponse> addStore(@Valid @RequestBody AddStoreRequest request) {
-        return new ResponseEntity<>(storeService.addStore(request), HttpStatus.CREATED);
+    public Long addStore(@Valid @RequestBody AddStoreRequest request) {
+        return storeService.addStore(request);
     }
 
     @PatchMapping("/update")
-    public ResponseEntity<StoreResponse> updateStore(@Valid @RequestBody UpdateStoreRequest request) {
-        return new ResponseEntity<>(storeService.updateStore(request), HttpStatus.CREATED);
+    public Long updateStore(@Valid @RequestBody UpdateStoreRequest request) {
+        return storeService.updateStore(request);
     }
 
-    @DeleteMapping
-    public ResponseEntity<Void> deleteStore(@RequestParam Long storeId) {
-        return new ResponseEntity<>(storeService.deleteStore(storeId),HttpStatus.OK);
+    @DeleteMapping("/{storeId}")
+    public void deleteStore(@PathVariable Long storeId) {
+        storeService.deleteStore(storeId);
     }
 
     @PostMapping("/open")
-    public String openStore() {
-        return storeService.openStore();
+    public void openStore(@PathVariable Long storeId) {
+        storeService.openStore(storeId);
     }
 
     @PostMapping("/close")
-    public String closeStore() {
-        return storeService.closeStore();
+    public void closeStore(@PathVariable Long storeId) {
+        storeService.closeStore(storeId);
     }
 }
